Rewrite lyapunov_exponent.m and make dynamical.m ergonomic
1. dynamical.m now returns accordingly based on the # of output params

2. The original code for lyapunov_exponent.m was taken from the slides. There
were several things that weren't clear in the code so I rewrote it from scratch.
For all values of a given parameters inside some range, we first calculate the
track and then after a transient period, takes the unit direction of dx and
calculates the divergence in that direction using the Jacobian. This
"n-dimensional" derivative is then averaged according to the lyapunov exponent
equation.