Finish 2d map and 3d flow lyapunov exponent calculation.
The best attractors for this analysis is yet to be determined. I removed
Lyapunov_exponents*.m and rewrote the plotting code inside lyapunov_exponents.m.
Generalize track_exp so that I can easily reuse it for the 3d case.

To calculate the exponents, you only need to calculate the trajectory and
jacobian with respect to a single varying parameter. I specialize the Jacobian
with this parameter and create the trajectory as usual.